# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Perf Assist Backend API
  version: 0.1.0

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /entries:
    get:
      summary: List entries by period
      operationId: listEntries
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
          required: false
          description: Start date (inclusive)
        - in: query
          name: to
          schema:
            type: string
            format: date
          required: false
          description: End date (inclusive)
      responses:
        '200':
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'

    post:
      summary: Create a new entry
      operationId: createEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntryRequest'
      responses:
        '201':
          description: Entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'

components:
  schemas:
    Entry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        type:
          type: string
          enum: [plan, fact]
        raw_text:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, user_id, date, type, raw_text, created_at]

    CreateEntryRequest:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        type:
          type: string
          enum: [plan, fact]
        raw_text:
          type: string
      required: [user_id, date, type, raw_text]